services/agent/requirements.txt
livekit-agents
livekit-api
python-dotenv
livekit
livekit-plugins-tavus
certifi
